apps/workflows-py/requirements.txt
dapr-ext-grpc>=1.17.0rc2
dapr-ext-workflow>=1.17.0rc2
dapr>=1.17.0rc2
flask>=2.3.0
opentelemetry-instrumentation-flask>=0.60b1
opentelemetry-instrumentation-grpc>=0.60b1
opentelemetry-sdk>=1.39.1
opentelemetry-exporter-otlp>=1.39.1
